Processes must be running on the server
