+++ libs/cgse-core/src/egse/procman/__init__.py
@@ -1,6 +1,8 @@
 import logging
+import subprocess
 import textwrap
 from pathlib import Path
+from typing import Union
 
 from egse.command import ClientServerCommand
 from egse.control import is_control_server_active
@@ -64,6 +66,133 @@ def get_status() -> str:
 
     else:
         return "Process Manager Status: [red]not active"
+
+
+class StartCommand:
+    """ Command to start the Control Server for a device."""
+
+    def __init__(self, device_id: str, cgse_cmd: str, device_args: Union[list, None], simulator_mode: bool = False):
+        """ Initialisation of a start command for a device Control Server.
+
+        Args:
+            device_id (str): Device identifier
+            cgse_cmd (str): CGSE command to start/stop the Control Server or to query its status
+            device_args (Union[list, None]): Device arguments
+            simulator_mode (bool): Whether to start the Control Server in simulator mode rather than operational mode
+        """
+
+        self._device_id = device_id
+        self._cgse_cmd = cgse_cmd
+        self._device_args = device_args
+        self._simulator_mode = simulator_mode
+
+    @property
+    def device_id(self) -> str:
+        """ Returns the device identifier.
+
+        Returns: Device identifier
+        """
+
+        return self._device_id
+
+    @property
+    def device_args(self) -> Union[list, None]:
+        """ Returns the device arguments.
+
+        Returns: Device arguments
+        """
+
+        return self._device_args
+
+    @property
+    def cmd(self) -> str:
+        """ Returns the full CGSE command to start the Control Server.
+
+        Returns: Full CGSE command to start the Control Server.
+        """
+
+        cmd = f"{self._cgse_cmd} start {self.device_id}"
+
+        if self.device_args:
+            cmd = f"{cmd} {self.device_args}"
+        if self.simulator_mode:
+            cmd = f"{cmd} --sim"
+
+        return cmd
+
+    @property
+    def simulator_mode(self) -> bool:
+        """ Checks whether the Control Server should be started in simulator mode rather than operational mode.
+
+        Returns: True if the Control Server should be started in simulator mode; False otherwise.
+        """
+
+        return self._simulator_mode
+
+
+class StopCommand:
+    """ Command to stop the Control Server for a device."""
+
+    def __init__(self, device_id: str, cgse_cmd: str):
+        """ Initialisation of a stop command for a device Control Server.
+
+        Args:
+            device_id (str): Device identifier
+            cgse_cmd (str): CGSE command to start/stop the Control Server or to query its status
+        """
+
+        self._device_id = device_id
+        self._cgse_cmd = cgse_cmd
+
+    @property
+    def device_id(self) -> str:
+        """ Returns the device identifier.
+
+        Returns: Device identifier
+        """
+
+        return self._device_id
+
+    @property
+    def cmd(self) -> str:
+        """ Returns the full CGSE command to stop the Control Server.
+
+        Returns: Full CGSE command to stop the Control Server.
+        """
+
+        return f"{self._cgse_cmd} stop {self.device_id}"
+
+
+class StatusCommand:
+    """ Command to query the status of a Control Server for a device."""
+
+    def __init__(self, device_id: str = None, cgse_cmd: str = None):
+        """ Initialisation of a status command for a device Control Server.
+
+        Args:
+            device_id (str): Device identifier
+            cgse_cmd (str): CGSE command to start/stop the Control Server or to query its status
+        """
+
+        self._device_id = device_id
+        self._cgse_cmd = cgse_cmd
+
+    @property
+    def device_id(self) -> str:
+        """ Returns the device identifier.
+
+        Returns: Device identifier
+        """
+
+        return self._device_id
+
+    @property
+    def cmd(self) -> str:
+        """ Returns the full CGSE command to query the status of the Control Server.
+
+        Returns: Full CGSE command to query the status of the Control Server.
+        """
+        return f"{self._cgse_cmd} status {self.device_id}"
 
 
 class ProcessManagerCommand(ClientServerCommand):
@@ -137,6 +266,37 @@ class ProcessManagerInterface(EventInterface):
 
         raise NotImplementedError
 
+    @dynamic_interface
+    def start_process(self, start_cmd: StartCommand) -> None:
+        """ Starts a process.
+
+        Args:
+            start_cmd (StartCommand): Command to start a process
+        """
+
+        raise NotImplementedError
+
+    @dynamic_interface
+    def stop_process(self, stop_cmd: StopCommand) -> None:
+        """ Stops a process.
+
+        Args:
+            stop_cmd (StopCommand): Command to stop a process
+        """
+
+        raise NotImplementedError
+
+    def get_process_status(self, status_cmd: StatusCommand) -> dict:
+        """ Returns the status of a process.
+
+        Args:
+            status_cmd (StatusCommand): Command to query the status of a process
+
+        Returns: Status of a process
+        """
+
+        raise NotImplementedError
+
 
 class ProcessManagerController(ProcessManagerInterface):
 
@@ -194,6 +354,13 @@ class ProcessManagerController(ProcessManagerInterface):
 
         self.setup = load_setup(setup_id=event.context["setup_id"])
 
+    def start_process(self, start_cmd: StartCommand):
+
+        subprocess.call(start_cmd.cmd, shell=True)
+
+    def stop_process(self, stop_cmd: StopCommand):
+
+        subprocess.call(stop_cmd.cmd, shell=True)
 
 class ProcessManagerProxy(Proxy, ProcessManagerInterface):
     """ Proxy for process management, used to connect to the Process Manager Control Server and send commands remotely.
